+++ packages/browser/src/transports/fetch.ts
@@ -133,23 +133,24 @@ export class FetchTransport extends BaseTransport {
     }
 
     return this._buffer.add(
-      new SyncPromise<Response>((resolve, reject) => {
-        void this._fetch(sentryRequest.url, options)
-          .then(response => {
-            const headers = {
-              'x-sentry-rate-limits': response.headers.get('X-Sentry-Rate-Limits'),
-              'retry-after': response.headers.get('Retry-After'),
-            };
-            this._handleResponse({
-              requestType: sentryRequest.type,
-              response,
-              headers,
-              resolve,
-              reject,
-            });
-          })
-          .catch(reject);
-      }),
+      () =>
+        new SyncPromise<Response>((resolve, reject) => {
+          void this._fetch(sentryRequest.url, options)
+            .then(response => {
+              const headers = {
+                'x-sentry-rate-limits': response.headers.get('X-Sentry-Rate-Limits'),
+                'retry-after': response.headers.get('Retry-After'),
+              };
+              this._handleResponse({
+                requestType: sentryRequest.type,
+                response,
+                headers,
+                resolve,
+                reject,
+              });
+            })
+            .catch(reject);
+        }),
     );
   }
 }